+++ History.md
@@ -1,3 +1,9 @@
+Next
+===================
+
+  * Fixed 'RangeError: Maximum call stack size exceeded' exception when parsing
+    over 10,000 nested css selectors
+
 v0.2.0 / 2014-11-24
 ===================